Do match Question (#44)
* do question

* do question controller

* add query filter for isDeleted feild

* add new permission: CHECK_QUIZ

* fix Do Question Function and add new quiz question function

* fix func doQuestion

* add require for Questiontype

* fix do match question

+++ QuizzBankBE.DataAccessLayer/DataObject/QuizResponse.cs
@@ -10,7 +10,7 @@ public partial class QuizResponse : IAuditedEntityBase
 
     public int? AccessId { get; set; }
 
-    public int? Mark { get; set; }
+    public float? Mark { get; set; }
 
     public string? Status { get; set; }
